+++ docs/changes.rst
@@ -22,6 +22,9 @@ Released: not yet
 
 **Cleanup:**
 
+* Extend use of general options in interactive mode to allow setting the
+  connections-file for an interactive command. (see issue #1037)
+
 **Known issues:**
 
 * See `list of open issues`_.